pyproject.toml
[project]
name = "revodata-gen-ai-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.3" 
dependencies = [
    "databricks-langchain>=0.4.2",
    "databricks-openai>=0.3.1",
    "databricks-sdk>=0.50.0",
    "databricks-vectorsearch>=0.55",
    "einops>=0.8.1",
    "langchain-huggingface>=0.1.2",
    "langgraph>=0.3.33",
    "mlflow[databricks]>=2.21.3",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.51.3",
]

[dependency-groups]
dev = [
    "databricks-agents>=0.19.0",
    "ipykernel>=6.29.5",
    "numpy>=1.26.4",
    "pre-commit>=4.2.0",
]
